kernel: fix -pedantic -pedantic-errors build errors
The Makefile now compiles with -pedantic -pedantic-errors. Adjust the code
to satisfy ISO C:

- install_ir(): pass an irq_handler_t function pointer instead of void*
- elf.h, multiboot2.h: give _Static_assert a message; replace zero-size and
  flexible-array members in packed structs with fixed-size arrays
- exception.c, proc.c, sched.c, qsort.c, syscall.c: route function-pointer
  <-> data-pointer conversions through uintptr_t; make start_kernel_proc
  and the syscall dispatch table use function-pointer/uintptr_t types
- assert.h: drop stray ';' after the inline function
- floppy.c: return a value from a void function
- vga.c: use ISO \033 instead of GNU \e ANSI escapes

+++ qsort.c
@@ -4,7 +4,7 @@
 
 static int compare_wrapper(const void* a, const void* b, void* arg)
 {
-	return ((int (*)(const void*, const void*)) arg)(a, b);
+	return ((int (*)(const void*, const void*)) (uintptr_t) arg)(a, b);
 }
 
 void qsort(void* base_ptr,
@@ -12,5 +12,6 @@ void qsort(void* base_ptr,
            size_t element_size,
            int (*compare)(const void*, const void*))
 {
-	qsort_r(base_ptr, num_elements, element_size, compare_wrapper, (void*) compare);
+	qsort_r(base_ptr, num_elements, element_size, compare_wrapper,
+	        (void *)(uintptr_t) compare);
 }
